Feat: top rated endpoint works
