--- app/controllers/movies_contoller.rb
@@ -1,25 +1,0 @@
-class MoviesController < ApplicationController
-    def top_rated
-        movies = MovieGateway.get_top_rated_movies
-        render json: format_response(movies)
-    end
-end
-
-private 
-
-def format_response(movies)
-    {
-        data: movies.first(20).map do |movie|
-            {
-                id: movie('id').to_s,
-                type: 'movie',
-                attributes: {
-                    title: movie['title']
-                    vote_average: movie['vote_average']
-
-                }
-            }
-        end
-    end
-    }
-end
